+++ src/main/java/com/bikeridediary/domain/bike/entity/BikeEntity.java
@@ -51,7 +51,7 @@ public class BikeEntity extends BaseEntity {
 
     // 현재 총 주행거리 (km) - 사용자가 수동으로 업데이트
     @Column(name = "total_mileage_km", nullable = false)
-    private Integer totalMileageKm;
+    private Long totalMileageKm;
 
     // 대표 바이크 여부 (정비/라이딩 기록의 기본값으로 사용)
     @Column(name = "is_representative", nullable = false)
@@ -81,7 +81,7 @@ public class BikeEntity extends BaseEntity {
             String modelName,
             Integer year,
             String category,
-            Integer totalMileageKm
+            Long totalMileageKm
     ) {
         BikeEntity bikeEntity = new BikeEntity();
         bikeEntity.userEntity = userEntity;
@@ -99,7 +99,7 @@ public class BikeEntity extends BaseEntity {
             String modelName,
             Integer year,
             String category,
-            Integer totalMileageKm,
+            Long totalMileageKm,
             LocalDate purchasedAt,
             String memo
     ) {
@@ -123,7 +123,7 @@ public class BikeEntity extends BaseEntity {
     }
 
     // 총 주행거리 갱신
-    public void updateMileage(int mileageKm) {
+    public void updateMileage(Long mileageKm) {
         this.totalMileageKm = mileageKm;
     }
