+++ src/main/java/com/bikeridediary/domain/fueling/dto/FuelingStatsResponse.java
@@ -15,6 +15,6 @@ public record FuelingStatsResponse(
         // 최근 연비 (km/L)
         BigDecimal latestFuelEfficiency,
         // 평균 리터당 가격 (원)
-        Integer averagePricePerLiter
+        Long averagePricePerLiter
 ) {
 }
