+++ src/main/java/com/bikeridediary/domain/maintenance/dto/MaintenanceScheduleUpdateRequest.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.*;
 public record MaintenanceScheduleUpdateRequest(
 
         @Min(value = 1, message = "km 기준 정비 주기는 1 이상이어야 합니다")
-        Integer intervalKm,
+        Long intervalKm,
 
         @Min(value = 1, message = "개월 기준 정비 주기는 1 이상이어야 합니다")
         Integer intervalMonths
