+++ src/main/java/com/bikeridediary/domain/maintenance/entity/MaintenanceScheduleEntity.java
@@ -39,7 +39,7 @@ public class MaintenanceScheduleEntity extends BaseEntity {
 
     // km 기준 정비 주기 (nullable - 없으면 km 기준 정비 안 함)
     @Column(name = "interval_km")
-    private Integer intervalKm;
+    private Long intervalKm;
 
     // 개월 기준 정비 주기 (nullable - 없으면 날짜 기준 정비 안 함)
     @Column(name = "interval_months")
@@ -49,7 +49,7 @@ public class MaintenanceScheduleEntity extends BaseEntity {
     public static MaintenanceScheduleEntity create(
             BikeEntity bikeEntity,
             MaintenanceType maintenanceType,
-            Integer intervalKm,
+            Long intervalKm,
             Integer intervalMonths
     ) {
         MaintenanceScheduleEntity entity = new MaintenanceScheduleEntity();
@@ -61,13 +61,13 @@ public class MaintenanceScheduleEntity extends BaseEntity {
     }
 
     // 정비 주기 정보 수정
-    public void update(Integer intervalKm, Integer intervalMonths) {
+    public void update(Long intervalKm, Integer intervalMonths) {
         this.intervalKm = intervalKm;
         this.intervalMonths = intervalMonths;
     }
 
     // km 기준으로 정비 필요 여부 확인 (마지막 정비 주행거리는 정비 기록에서 전달)
-    public boolean isOverdueByKm(Integer currentMileage, Integer lastMaintenanceMileage) {
+    public boolean isOverdueByKm(Long currentMileage, Long lastMaintenanceMileage) {
         if (intervalKm == null || lastMaintenanceMileage == null) {
             return false;
         }
@@ -83,8 +83,8 @@ public class MaintenanceScheduleEntity extends BaseEntity {
     }
 
     // km 또는 날짜 중 하나라도 정비 주기를 초과했으면 true 반환
-    public boolean isOverdue(Integer currentMileage, LocalDate currentDate,
-                             Integer lastMaintenanceMileage, LocalDate lastMaintenanceDate) {
+    public boolean isOverdue(Long currentMileage, LocalDate currentDate,
+                             Long lastMaintenanceMileage, LocalDate lastMaintenanceDate) {
         return isOverdueByKm(currentMileage, lastMaintenanceMileage)
                 || isOverdueByDate(currentDate, lastMaintenanceDate);
     }
